% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summary.R
\name{base_summary}
\alias{base_summary}
\title{Statistical summary using mostly base functions}
\usage{
base_summary(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a data.frame with summary statistics
}
\description{
Statistical summary using mostly base functions
}
\examples{
base_summary(mtcars)
}
