% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_fun.R
\name{qual_fun}
\alias{qual_fun}
\title{Quality function for dataset using base functions}
\usage{
qual_fun(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a data frame with the number of NA, NaN and infinite values per variables.
}
\description{
Quality function for dataset using base functions
}
\examples{
qual_fun(iris)
}
